Restyle subscribe screen as a standard dialog and align report dialog layout
